refactor: #40 eslint 경고해결
