+++ src/hooks/useKakaoMap.js
@@ -131,8 +131,8 @@ const useKakaoMap = (appKey, initialLocation, isBoring = false) => {
           );
           const dogMarkerImage = new window.kakao.maps.MarkerImage(
             dog.image,
-            new window.kakao.maps.Size(50, 50),
-            { shape: 'rect', coords: '0,0,50,50' }
+            new window.kakao.maps.Size(20, 20),
+            { shape: "rect", coords: "0,0,50,50" },
           );
 
           const dogMarker = new window.kakao.maps.Marker({
@@ -223,4 +223,3 @@ const useKakaoMap = (appKey, initialLocation, isBoring = false) => {
 };
 
 export default useKakaoMap;
-
